FoodVISION_VIT_FINALLY/requirements.txt
gradio>=4.31.0
torch
torchvision
timm
Pillow
matplotlib